requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==45.0.7
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gcloud==0.18.3
googleapis-common-protos==1.70.0
httplib2==0.30.0
idna==3.10
jwcrypto==1.5.6
oauth2client==4.1.3
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
